requirements.txt
streamlit
transformers
bitsandbytes
bitsandbytes-windows
reportlab
joblib
sentence-transformers
tqdm
torch==2.2.0
pandas
pypdf
accelerate
Flask
Pandas
Numpy
faiss-cpu